requirements.txt
python-dotenv
# Had to specify an older version to avoid clashing with cryptography libraries
# jwt 1.3.1 and pyjwt 2.6.0, 2.5.0, 2.3.0, 2.2.0 cause:
# jwt===1.3.0 is verified to work, but I had intended to import pyjwt
# installed: not sure if this will be an issue in the ros2 branch of the rosbridge suite
# File "/usr/lib/python3/dist-packages/OpenSSL/crypto.py", line 1573, in X509StoreFlags
#     CB_ISSUER_CHECK = _lib.X509_V_FLAG_CB_ISSUER_CHECK
# AttributeError: module 'lib' has no attribute 'X509_V_FLAG_CB_ISSUER_CHECK'
# [INFO] [1673988598.375001]: Rosapi started
pyjwt[crypto]===2.1.0